Use identifier in time-series data (#37)

+++ homeserver-common/src/main/java/org/juhanir/domain/sensordata/dto/outgoing/HumidityStatusResponse.java
@@ -16,7 +16,7 @@ public record HumidityStatusResponse(
     public static HumidityStatusResponse fromHumidityStatus(HumidityStatus humidStatus) {
         return new HumidityStatusResponse(
                 humidStatus.getComponentId(),
-                humidStatus.getValue().doubleValue(),
+                humidStatus.getValue(),
                 humidStatus.getMeasurementTime(),
                 humidStatus.getDevice().getIdentifier()
         );
